backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Redis & Caching
redis==5.0.1
celery==5.3.4

# AWS & Authentication
boto3==1.29.7
pyjwt==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# ML/NLP
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2

# Security
bcrypt==4.1.1
python-multipart==0.0.6

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.38.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0
